Add support for partial downloads
Currently, files may only be downloaded from start to finish. However,
it is common for applications to want to request only part of a file.
For example, video players which allow seeking.

To support this, take startPos and endPos arguments in the
DownloadAsync method. This is an inclusive byte range, so requesting
0-499 for example will return the first 500 bytes of the file. This is
consistent with how the HTTP Range header works.