Switch to per-file download concurrency limit
Currently, at most 8 blocks may be downloaded concurrently within an
application. However, even blocks which have finished downloading and
are waiting to be processed count towards this limit. Which means that a
single file download with slow processing could block all other
downloads. (A common scenario in video players for example, which only
process blocks as fast as the video is being played.)

To fix this, remove the global block download concurrency limit and
switch to a per-file limit.